+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxCert.hs
@@ -77,10 +77,8 @@ module Cardano.Ledger.Shelley.TxCert (
 import Cardano.Ledger.BaseTypes (invalidKey, kindObject)
 import Cardano.Ledger.Binary (
   DecCBOR (decCBOR),
-  DecCBORGroup (..),
   Decoder,
   EncCBOR (..),
-  EncCBORGroup (..),
   Encoding,
   FromCBOR (..),
   ToCBOR (..),
@@ -90,7 +88,6 @@ import Cardano.Ledger.Binary (
   decodeWord,
   encodeListLen,
   encodeWord8,
-  listLenInt,
   peekTokenType,
  )
 import Cardano.Ledger.Coin (Coin (..), DeltaCoin)
@@ -105,7 +102,12 @@ import Cardano.Ledger.Internal.Era (AllegraEra, AlonzoEra, BabbageEra, MaryEra)
 import Cardano.Ledger.Keys (asWitness)
 import Cardano.Ledger.Shelley.Era (ShelleyEra)
 import Cardano.Ledger.Shelley.PParams ()
-import Cardano.Ledger.State (PoolParams (..))
+import Cardano.Ledger.State (
+  PoolParams (..),
+  decCBORGroupPoolParams,
+  encCBORGroupPoolParams,
+  poolParamsCount,
+ )
 import Cardano.Ledger.Val ((<+>), (<×>))
 import Control.DeepSeq (NFData (..), rwhnf)
 import Data.Aeson (ToJSON (..), (.=))
@@ -434,9 +436,9 @@ encodeShelleyDelegCert = \case
 encodePoolCert :: PoolCert -> Encoding
 encodePoolCert = \case
   RegPool poolParams ->
-    encodeListLen (1 + listLen poolParams)
+    encodeListLen (1 + fromIntegral poolParamsCount)
       <> encodeWord8 3
-      <> encCBORGroup poolParams
+      <> encCBORGroupPoolParams poolParams
   RetirePool vk epoch ->
     encodeListLen 3
       <> encodeWord8 4
@@ -504,8 +506,8 @@ shelleyTxCertDelegDecoder = \case
 poolTxCertDecoder :: EraTxCert era => Word -> Decoder s (Int, TxCert era)
 poolTxCertDecoder = \case
   3 -> do
-    group <- decCBORGroup
-    pure (1 + listLenInt group, RegPoolTxCert group)
+    group <- decCBORGroupPoolParams
+    pure (1 + fromIntegral poolParamsCount, RegPoolTxCert group)
   4 -> do
     a <- decCBOR
     b <- decCBOR